restore old index.tsx correctly
